requirements.txt
Flask==3.1.0
python-dotenv==0.21.0
psycopg2-binary==2.9.3
Flask-SQLAlchemy==3.1.1